Commit 3: Fase 2 completa (implementacion de parseFloat)

+++ script.js
@@ -1,7 +1,7 @@
 /*Archivo de JavaScript*/
 
-let a = parseInt(prompt("Ingrese el primer número:"));
-let b = parseInt(prompt("Ingrese el segundo número:"));
+let a = parseFloat(prompt("Ingrese el primer número:"));
+let b = parseFloat(prompt("Ingrese el segundo número:"));
 
 //Funciones de operaaciones 
 function sumar(a, b) {
